VS Code extension with Preview
Preview with auto-update while editing.  This version uses the watch feature of d2 to auto-update the browser view in VS Code.

To make this work, the following change needs to be made in the 'd2' project:

diff --git a/watch.go b/watch.go
index 60b8e90..555849a 100644
--- a/watch.go
+++ b/watch.go
@@ -447,7 +447,8 @@ func (w *watcher) handleWatch(hw http.ResponseWriter, r *http.Request) error {
        w.wsclientsMu.Unlock()

        c, err := websocket.Accept(hw, r, &websocket.AcceptOptions{
-               CompressionMode: websocket.CompressionDisabled,
+               CompressionMode:    websocket.CompressionDisabled,
+               InsecureSkipVerify: true,
        })
        if err != nil {
                w.wsclientsWG.Done()
(END)

This turns off cross-site scripting, which is beyond my pay grade to determine if this is ok.

If this is not ok, then I can make this work, but with more changes to 'd2'