Now with more colors...
Added theming to the webView so the toolbar matches vscode theme.
Added syntax coloring values to package.json
Changed extension load event to onLanguage:d2 instead of onStartupFinished to be a better extension citizen.
Changed formatting feature to use a formatting provider, so format on exit works automagically and is initiated by the standard 'format document' command instead of a specific format command for d2 files.